Split the root layer stack out
It's a fairly complex bit of code, so it deserves its own module.

Update tests.
