+++ lib/layer/__init__.py
@@ -32,4 +32,4 @@ from .consts import *
 from .group import *
 from .data import *
 from .core import *
-
+from .tree import *
